move logic process to service layer(For both event and post)
