+++ api/v1/post.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
+import service.post
 import repository.post
 from schema.database.postupdate import PostUpdate
 from infrastructure.mysql import get_db
@@ -25,21 +26,21 @@ def create_post(post: PostCreate, db: Session = Depends(get_db)):
 
 @router.delete("/{post_id}")
 def delete(post_id: int, db: Session = Depends(get_db)):
-    post = repository.post.get_post(db, post_id)
-    return repository.post.delete(db=db, post=post)
+    post = service.post.get_post(db, post_id)
+    return service.post.delete(db=db, post=post)
 
 
 @router.patch("/{post_id}")
 def update_post(post_id: int, post_update: PostUpdate, db: Session = Depends(get_db)):
-    existing_post = repository.post.get_post(db, post_id)
+    existing_post = service.post.get_post(db, post_id)
     if existing_post is None:
         raise HTTPException(status_code=404, detail="Post not found")
-    updated_post = repository.post.patch_post(db, post_id, post_update)
+    updated_post = service.post.patch_post(db, post_id, post_update)
     return updated_post
 
 @router.get("/{post_id}")
 def get_post(post_id: int, db: Session = Depends(get_db)):
-    post = repository.post.get_post(db, post_id)
+    post = service.post.get_post(db, post_id)
     if post is None:
         raise HTTPException(status_code=404, detail="Post not found")
     return post
